+++ cluster-controller/src/main/java/io/strimzi/controller/cluster/ClusterController.java
@@ -10,10 +10,10 @@ import io.fabric8.kubernetes.client.KubernetesClientException;
 import io.fabric8.kubernetes.client.Watch;
 import io.fabric8.kubernetes.client.Watcher;
 import io.netty.handler.codec.http.HttpResponseStatus;
-import io.strimzi.controller.cluster.operator.cluster.AbstractClusterOperations;
-import io.strimzi.controller.cluster.operator.cluster.KafkaClusterOperations;
-import io.strimzi.controller.cluster.operator.cluster.KafkaConnectClusterOperations;
-import io.strimzi.controller.cluster.operator.cluster.KafkaConnectS2IClusterOperations;
+import io.strimzi.controller.cluster.operator.assembly.AbstractClusterOperations;
+import io.strimzi.controller.cluster.operator.assembly.KafkaClusterOperations;
+import io.strimzi.controller.cluster.operator.assembly.KafkaConnectClusterOperations;
+import io.strimzi.controller.cluster.operator.assembly.KafkaConnectS2IClusterOperations;
 import io.strimzi.controller.cluster.resources.KafkaCluster;
 import io.strimzi.controller.cluster.resources.KafkaConnectCluster;
 import io.strimzi.controller.cluster.resources.KafkaConnectS2ICluster;
