+++ cluster-controller/src/main/java/io/strimzi/controller/cluster/Main.java
@@ -8,11 +8,11 @@ import io.fabric8.kubernetes.client.DefaultKubernetesClient;
 import io.fabric8.kubernetes.client.KubernetesClient;
 import io.fabric8.openshift.client.OpenShiftClient;
 import io.netty.handler.codec.http.HttpResponseStatus;
-import io.strimzi.controller.cluster.operator.cluster.KafkaClusterOperations;
-import io.strimzi.controller.cluster.operator.cluster.KafkaConnectClusterOperations;
-import io.strimzi.controller.cluster.operator.cluster.KafkaConnectS2IClusterOperations;
-import io.strimzi.controller.cluster.operator.cluster.KafkaSetOperations;
-import io.strimzi.controller.cluster.operator.cluster.ZookeeperSetOperations;
+import io.strimzi.controller.cluster.operator.assembly.KafkaClusterOperations;
+import io.strimzi.controller.cluster.operator.assembly.KafkaConnectClusterOperations;
+import io.strimzi.controller.cluster.operator.assembly.KafkaConnectS2IClusterOperations;
+import io.strimzi.controller.cluster.operator.assembly.KafkaSetOperations;
+import io.strimzi.controller.cluster.operator.assembly.ZookeeperSetOperations;
 import io.strimzi.controller.cluster.operator.resource.BuildConfigOperations;
 import io.strimzi.controller.cluster.operator.resource.ConfigMapOperations;
 import io.strimzi.controller.cluster.operator.resource.DeploymentConfigOperations;
