+++ cluster-controller/src/main/java/io/strimzi/controller/cluster/resources/Labels.java
@@ -35,11 +35,11 @@ public class Labels {
     public static final String STRIMZI_KIND_LABEL = STRIMZI_DOMAIN + "kind";
     /**
      * The type of Strimzi component:
-     * E.g: {@link io.strimzi.controller.cluster.operator.cluster.KafkaClusterOperations#CLUSTER_TYPE_KAFKA kafka},
-     * {@link io.strimzi.controller.cluster.operator.cluster.KafkaClusterOperations#CLUSTER_TYPE_ZOOKEEPER zookeeper},
-     * {@link io.strimzi.controller.cluster.operator.cluster.KafkaClusterOperations#CLUSTER_TYPE_TOPIC_CONTROLLER topic-controller},
-     * {@link io.strimzi.controller.cluster.operator.cluster.KafkaConnectClusterOperations#CLUSTER_TYPE_CONNECT kafka-connect},
-     * {@link io.strimzi.controller.cluster.operator.cluster.KafkaConnectS2IClusterOperations#CLUSTER_TYPE_CONNECT_S2I kafka-connect-s2i}
+     * E.g: {@link io.strimzi.controller.cluster.operator.assembly.KafkaClusterOperations#CLUSTER_TYPE_KAFKA kafka},
+     * {@link io.strimzi.controller.cluster.operator.assembly.KafkaClusterOperations#CLUSTER_TYPE_ZOOKEEPER zookeeper},
+     * {@link io.strimzi.controller.cluster.operator.assembly.KafkaClusterOperations#CLUSTER_TYPE_TOPIC_CONTROLLER topic-controller},
+     * {@link io.strimzi.controller.cluster.operator.assembly.KafkaConnectClusterOperations#CLUSTER_TYPE_CONNECT kafka-connect},
+     * {@link io.strimzi.controller.cluster.operator.assembly.KafkaConnectS2IClusterOperations#CLUSTER_TYPE_CONNECT_S2I kafka-connect-s2i}
      */
     public static final String STRIMZI_TYPE_LABEL = STRIMZI_DOMAIN + "type";
